feat(analytics)!: extract @granit/react-ui-analytics; react-analytics is now react-ui-free
Option-b strict tiers: the shadcn-styled analytics widget config forms
(kpi / chart / table / pivot) and the shared query-binding editor controls
move out of the headless @granit/react-analytics into a new react-ui tier
package, @granit/react-ui-analytics. react-analytics keeps only headless
building blocks: hooks, unstyled tiles, the widget registry, the widget
catalog, required-field validation, and the useQueryFieldMetadata hook.

The former query-field-controls.tsx is split along the seam: the headless
useQueryFieldMetadata hook + FieldOption/QueryFieldMetadata types stay in
react-analytics (editor/use-query-field-metadata.ts, re-exported from the
root and /editor barrels); the shadcn controls move to react-ui-analytics.

@granit/react-ui is dropped from react-analytics peer/devDeps. Consumers
rewired: react-ui-dashboards (dashboard composer) and react-map (map widget
config form) now import the forms/controls/registry from
@granit/react-ui-analytics and gained it as a (peer)dependency.

BREAKING CHANGE: the analytics widget config forms and controls no longer
export from @granit/react-analytics/editor. Import them from the new
@granit/react-ui-analytics package instead:
  - Moved: TableConfigForm, KpiConfigForm, ChartConfigForm, PivotConfigForm,
    analyticsWidgetConfigFormRegistry, EnumSelect, MetaFieldInput,
    MetaMultiFieldInput, QueryNameCombobox, RequiredMark.
  - Still in @granit/react-analytics/editor: analyticsWidgetCatalog,
    validateWidgetConfig, isWidgetConfigComplete, useQueryFieldMetadata,
    FieldOption, QueryFieldMetadata.
A coordinated showcase-admin-react MR is required.
