+++ packages/@granit/react-analytics/package.json
@@ -23,7 +23,6 @@
     "@granit/react-dashboard-editor": "workspace:*",
     "@granit/react-dashboards": "workspace:*",
     "@granit/react-query-engine": "workspace:*",
-    "@granit/react-ui": "workspace:*",
     "@granit/query-engine": "workspace:*",
     "@granit/types": "workspace:*",
     "@tanstack/react-query": "^5.0.0",
@@ -67,7 +66,6 @@
     "@granit/react-dashboards": "workspace:*",
     "@granit/react-query-engine": "workspace:*",
     "@granit/react-testing": "workspace:*",
-    "@granit/react-ui": "workspace:*",
     "@granit/query-engine": "workspace:*",
     "@granit/testing": "workspace:*",
     "@granit/types": "workspace:*",
