+++ packages/@granit/react-map/package.json
@@ -22,6 +22,7 @@
     "@granit/react-analytics": "workspace:*",
     "@granit/react-dashboard-editor": "workspace:*",
     "@granit/react-dashboards": "workspace:*",
+    "@granit/react-ui-analytics": "workspace:*",
     "@granit/utils": "workspace:*",
     "@tanstack/react-query": "^5.0.0",
     "dompurify": "^3.0.0",
@@ -32,6 +33,9 @@
   },
   "peerDependenciesMeta": {
     "@granit/react-analytics": {
+      "optional": true
+    },
+    "@granit/react-ui-analytics": {
       "optional": true
     },
     "@tanstack/react-query": {
@@ -75,6 +79,7 @@
     "@granit/react-dashboard-editor": "workspace:*",
     "@granit/react-dashboards": "workspace:*",
     "@granit/react-testing": "workspace:*",
+    "@granit/react-ui-analytics": "workspace:*",
     "@granit/testing": "workspace:*",
     "@granit/utils": "workspace:*",
     "@tanstack/react-query": "^5.0.0",
