+++ packages/@granit/react-ui-analytics/package.json
@@ -1,0 +1,48 @@
+{
+  "name": "@granit/react-ui-analytics",
+  "description": "shadcn-styled config forms for the analytics dashboard widgets (kpi / chart / table / pivot) plus the shared query-binding editor controls. Composes @granit/react-analytics (headless hooks + widget metadata) with the foundation UI packages for the dashboard composer.",
+  "version": "0.1.0",
+  "license": "Apache-2.0",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsup"
+  },
+  "peerDependencies": {
+    "@granit/analytics": "workspace:*",
+    "@granit/dashboards": "workspace:*",
+    "@granit/query-engine": "workspace:*",
+    "@granit/react-analytics": "workspace:*",
+    "@granit/react-dashboard-editor": "workspace:*",
+    "@granit/react-ui": "workspace:*",
+    "i18next": "^26.0.0",
+    "react": "^19.0.0",
+    "react-i18next": "^17.0.0"
+  },
+  "publishConfig": {
+    "exports": {
+      ".": {
+        "types": "./dist/index.d.ts",
+        "import": "./dist/index.js"
+      }
+    },
+    "registry": "https://npm.pkg.github.com"
+  },
+  "devDependencies": {
+    "@granit/analytics": "workspace:*",
+    "@granit/api-client": "workspace:*",
+    "@granit/dashboards": "workspace:*",
+    "@granit/query-engine": "workspace:*",
+    "@granit/react-analytics": "workspace:*",
+    "@granit/react-dashboard-editor": "workspace:*",
+    "@granit/react-query-engine": "workspace:*",
+    "@granit/react-testing": "workspace:*",
+    "@granit/react-ui": "workspace:*",
+    "i18next": "^26.0.0"
+  }
+}
