+++ packages/@granit/react-ui-analytics/src/components/chart-config-form.tsx
@@ -1,3 +1,4 @@
+import { useQueryFieldMetadata } from '@granit/react-analytics';
 import { useTranslation } from 'react-i18next';
 
 import {
@@ -5,7 +6,6 @@ import {
   MetaFieldInput,
   QueryNameCombobox,
   RequiredMark,
-  useQueryFieldMetadata,
 } from './query-field-controls';
 
 import type { ChartType, ChartWidgetDefinition } from '@granit/analytics';
