+++ packages/@granit/react-ui-analytics/src/components/query-field-controls.tsx
@@ -1,15 +1,15 @@
 // ---------------------------------------------------------------------------
 // Shared editor controls — catalogue-backed query picker + metadata-backed
-// field selectors used by the chart / table / pivot config forms.
+// field selectors used by the chart / table / pivot / kpi config forms.
 //
 // Built on the shadcn Combobox / Select from @granit/react-ui. The comboboxes
 // accept a typed value outside the option list (allowCustomValue), so they keep
 // working as free-text-with-suggestions when no <QueryCatalogProvider> is present
 // or the selected query has no resolvable metadata. A <QueryClientProvider> is
-// required for the catalogue/metadata fetches.
+// required for the catalogue/metadata fetches (see `useQueryFieldMetadata` in
+// @granit/react-analytics).
 // ---------------------------------------------------------------------------
 
-import { useQueryCatalog, useQueryMetaAt } from '@granit/react-query-engine';
 import {
   Combobox,
   ComboboxMulti,
@@ -23,85 +23,8 @@ import {
 import { useTranslation } from 'react-i18next';
 
 import type { QueryCatalogEntryResponse } from '@granit/query-engine';
+import type { FieldOption } from '@granit/react-analytics';
 import type { TFunction } from 'i18next';
-
-/** CLR type names eligible for numeric aggregation (Sum/Avg/Min/Max). */
-const NUMERIC_CLR_TYPES = new Set([
-  'Byte',
-  'SByte',
-  'Int16',
-  'UInt16',
-  'Int32',
-  'UInt32',
-  'Int64',
-  'UInt64',
-  'Single',
-  'Double',
-  'Decimal',
-]);
-
-export interface FieldOption {
-  readonly name: string;
-  readonly label?: string;
-}
-
-export interface QueryFieldMetadata {
-  /** Catalogue entries for the query combobox, or `undefined` with no provider. */
-  readonly catalogEntries: readonly QueryCatalogEntryResponse[] | undefined;
-  readonly hasCatalog: boolean;
-  /**
-   * Group By dimension options: the query's declared group-by fields, or — when
-   * a query declares none — every column as a fallback so the control still
-   * offers a real picker.
-   */
-  readonly groupByOptions: readonly FieldOption[];
-  /**
-   * Field (aggregation target) options: numeric columns, or — when none are
-   * detected (e.g. nullable numerics the backend reports as `Nullable`1`) —
-   * every column as a fallback.
-   */
-  readonly fieldOptions: readonly FieldOption[];
-  /** All columns of the selected query (e.g. the table visible-columns picker). */
-  readonly columnOptions: readonly FieldOption[];
-}
-
-/**
- * Resolves the query catalogue and the selected query's metadata into the
- * field-option lists the config forms render. Returns empty option lists until
- * a catalogue entry matching `queryName` resolves its metadata; once columns
- * load, Group By and Field always have options (falling back to all columns).
- */
-export function useQueryFieldMetadata(queryName: string): QueryFieldMetadata {
-  const { data: catalog } = useQueryCatalog();
-  const selectedEntry = catalog?.find((entry) => entry.name === queryName) ?? null;
-  const { data: meta } = useQueryMetaAt(selectedEntry?.basePath ?? null);
-
-  const toOption = (field: { name: string; label?: string }): FieldOption => ({
-    name: field.name,
-    label: field.label,
-  });
-  const columnOptions = (meta?.columns ?? []).map(toOption);
-  const groupByDeclared = (meta?.groupByFields ?? []).map(toOption);
-  const numericColumns = (meta?.columns ?? [])
-    .filter((column) => NUMERIC_CLR_TYPES.has(column.type))
-    .map(toOption);
-
-  return {
-    catalogEntries: catalog,
-    hasCatalog: catalog != null && catalog.length > 0,
-    groupByOptions: groupByDeclared.length > 0 ? groupByDeclared : columnOptions,
-    fieldOptions: numericColumns.length > 0 ? numericColumns : columnOptions,
-    columnOptions,
-  };
-}
-
-/** Splits a comma-separated free-text field list into trimmed, non-empty names. */
-export function splitFields(raw: string): string[] {
-  return raw
-    .split(',')
-    .map((segment) => segment.trim())
-    .filter((segment) => segment.length > 0);
-}
 
 const toComboboxOptions = (options: readonly FieldOption[]): ComboboxOption[] =>
   options.map((option) => ({ value: option.name, label: option.label }));
