+++ packages/@granit/react-ui-dashboards/package.json
@@ -23,6 +23,7 @@
     "@granit/react-localization": "workspace:*",
     "@granit/react-map": "workspace:*",
     "@granit/react-ui": "workspace:*",
+    "@granit/react-ui-analytics": "workspace:*",
     "@granit/react-ui-kit": "workspace:*",
     "lucide-react": "^1.21.0",
     "react": "^19.0.0",
@@ -50,6 +51,7 @@
     "@granit/react-localization": "workspace:*",
     "@granit/react-map": "workspace:*",
     "@granit/react-ui": "workspace:*",
+    "@granit/react-ui-analytics": "workspace:*",
     "@granit/react-ui-kit": "workspace:*",
     "@storybook/react-vite": "^10.4.6",
     "@tanstack/react-query": "^5.0.0",
